# docs/swagger.yaml
basePath: /
definitions:
  dto.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  dto.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.APIError'
      request_id:
        type: string
      success:
        type: boolean
    type: object
...
  dto.JWTDecodeRequest:
    properties:
      secret:
        description: Secret is optional. When provided the signature is verified.
        type: string
      token:
        description: Token is the full JWT string (header.payload.signature).
        type: string
    required:
    - token
    type: object
  dto.JWTDecodeResponse:
    properties:
      error:
        type: string
      expires_at:
        type: string
      header:
        additionalProperties: true
        type: object
      issued_at:
        type: string
      not_before:
        type: string
      payload:
        additionalProperties: true
        type: object
      signature:
        type: string
      valid:
        type: boolean
    type: object
  dto.JWTGenerateRequest:
    properties:
      claims:
        additionalProperties: true
        description: Claims are the custom payload claims to embed in the token.
        type: object
      expires_in_seconds:
        description: ExpiresInSeconds is the token TTL in seconds. Defaults to 3600
          if zero.
        example: 3600
        type: integer
      secret:
        description: Secret is the HMAC-SHA256 signing key.
        example: my-super-secret
        type: string
    required:
    - claims
    - secret
    type: object
  dto.JWTGenerateResponse:
    properties:
      token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A stateless developer toolbox providing HTTP inspection, cURL conversion,
    JWT toolkit, and JSON generation.
  title: DevHelp API
  version: 1.0.0
paths:
...
  /api/v1/jwt/decode:
    post:
      consumes:
      - application/json
      description: Decodes a JWT and returns header, payload, signature, and time
        claims. Optionally verifies the signature.
      parameters:
      - description: JWT decode input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JWTDecodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JWTDecodeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Decode a JWT
      tags:
      - JWT Toolkit
  /api/v1/jwt/generate:
    post:
      consumes:
      - application/json
      description: Signs and returns a new HS256 JWT with the provided claims and
        secret.
      parameters:
      - description: JWT generation input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JWTGenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.JWTGenerateResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Generate a JWT
      tags:
      - JWT Toolkit
schemes:
- http
- https
swagger: "2.0"